finish yscd

+++ src/views/ysdaComponents/yscd.vue
@@ -6,12 +6,7 @@
     <el-row style="margin-top: 40px">
       <el-table :data="initFormData" border stripe max-height="750">
         <el-table-column type="index" label="序号" />
-        <el-table-column label="操作">
-          <template slot-scope="scope">
-            <el-button type="text" @click="editData(scope.row.id)">修改</el-button>
-            <el-button type="text" @click="deleteData(scope.row.id)">删除</el-button>
-          </template>
-        </el-table-column>
+
         <el-table-column label="产地" prop="chanDi" />
         <el-table-column label="预计到货天数" prop="yjdhts" />
         <el-table-column label="耗纱系数" prop="hsxs" />
@@ -21,6 +16,12 @@
           </template>
         </el-table-column>
         <el-table-column label="备注" prop="bz" />
+        <el-table-column label="操作">
+          <template slot-scope="scope">
+            <el-button type="text" @click="editData(scope.row.id)">修改</el-button>
+            <el-button type="text" @click="deleteData(scope.row.id)">删除</el-button>
+          </template>
+        </el-table-column>
       </el-table>
     </el-row>
     <el-dialog v-if="dialogEditVisible" title="原纱产地（编辑）" :visible.sync="dialogEditVisible" width="95%">
@@ -60,15 +61,19 @@ export default {
       })
     },
     deleteData(id) {
-      var url = '/api/deleteYarnChanDi/' + id
-      deleteChanDi(url).then(res => {
-        if (res.status !== 200) {
-          this.$message.error(res.data.tipInfo)
-        } else {
-          this.$message.success(res.data.tipInfo)
-          this.initData()
-        }
-      })
+      this.$confirm('确认删除？')
+        .then(_ => {
+          var url = '/api/deleteYarnChanDi/' + id
+          deleteChanDi(url).then(res => {
+            if (res.status !== 200) {
+              this.$message.error(res.data.tipInfo)
+            } else {
+              this.$message.success(res.data.tipInfo)
+              this.initData()
+            }
+          })
+        })
+        .catch(_ => {})
     },
     editData(id) {
       var url = '/api/getOneYarnChanDi/' + id
